Split NetCore and NetStandard Platform Proxy classes.
