+++ src/client/Microsoft.Identity.Client.Services/Platforms/netstandard/NetstandardPlatformProxy.cs
